+++ headers/umount_tab.go
@@ -1,7 +1,7 @@
 package headers
 
 // Location: include/linux/fs.h
-var UmountLookUp = map[int]string{
+var UmountLookUp = map[int64]string{
 	0x00000001: "MNT_FORCE",
 	0x00000002: "MNT_DETACH",
 	0x00000004: "MNT_EXPIRE",
